A bunch of changes
Still no rate page but the structure is ready and accessible by two methods. To see them load the page in the browser and open the console window. 
Then type - app.form.convertToData('#myForm') to see the data as it is in the form in local storage or bodySystems to see the global variable.