Severity Page
re-created page to rate symptom severity. You have to have a server to run it locally. Feel free to download mongoose, but I'll merge the changes with Master so that you all can view it on gitHub.
I still need to add event lisenters to pass data and update the previous screen. We also need to decide if we want to add a text area for comments.